backend/api/requirements.txt
fastapi==0.110.1
mangum==0.17.0
firebase-admin>=6.2.0
pydantic>=2.6.4
python-dotenv>=1.0.1 